+++ pytorch_lightning/trainer/callback_hook.py
@@ -232,7 +232,12 @@ class TrainerCallbackHookMixin(ABC):
             callback.on_predict_end(self, self.lightning_module)
 
     def on_keyboard_interrupt(self):
-        """Called when the training is interrupted by KeyboardInterrupt."""
+        r"""
+        .. deprecated:: v1.5
+            This callback hook was deprecated in v1.5 in favor of `on_exception` and will be removed in v1.7.
+
+        Called when any trainer execution is interrupted by KeyboardInterrupt.
+        """
         for callback in self.callbacks:
             callback.on_keyboard_interrupt(self, self.lightning_module)